backend/python_backend/requirements.txt
flask
joblib
pandas
scikit-learn
lime
numpy
pydantic  # ✨ NEW: For data validation
python-dotenv  # ✨ NEW: For .env file support